Se modifica el nombre de la clase Controladora de excepciones de la api Se modifica los get y list de generos y se habilita el post con una estructura de excepciones try y catch.  hay que crear un sp para  ahorrar id y manipular los insert y update
