+++ src/main/java/com/javadabadu/disney/models/dto/ResponseInfoDTO.java
@@ -1,19 +1,20 @@
 package com.javadabadu.disney.models.dto;
 
 import lombok.Data;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 
 import java.time.LocalDate;
 
 @Data
+@RequiredArgsConstructor
 public class ResponseInfoDTO {
     private String timestap = LocalDate.now().toString();
+    @NonNull
     private String message;
+    @NonNull
     private String path;
+    @NonNull
     private Integer codeStatus;
 
-    public ResponseInfoDTO(String message, String path, Integer codeStatus) {
-        this.message = message;
-        this.path = path;
-        this.codeStatus = codeStatus;
-    }
 }
